+++ CHANGES.md
@@ -1,10 +1,14 @@
 
 ### Version 1.4.12
 
-Element assembly for calculation of current and charges from the device into the circuit are fixed.  These examples are added to the tests:
+Element assembly for calculation of current and charges from the device into the circuit equation are fixed.  These tests are added:
 
 - ``testing/cap_2d_edge.py``
 - ``testing/cap_2d_element.py``
+- ``testing/cap_3d_edge.py``
+- ``testing/cap_3d_element.py``
+
+The ``edge`` variant is using standard edge based assembly, and the ``element`` variant is using element-based assembly.
 
 ### Version 1.4.11